Added buttonsWithReply | test 4
